+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Fixed
 UI
+- Text input via IME was broken
 - Double clicking in text field may lose focus
 
 ## UI [2.0.1] - 2025-01-25